hepatic: num workers 28

+++ heart_experiment/metrics/metrics.py
@@ -1,5 +1,5 @@
 from sklearn.metrics import multilabel_confusion_matrix
-
+import numpy as np
 
 class Metric:
     def __init__(self):
@@ -9,10 +9,10 @@ class Metric:
     def calc_cm(self, labels, outputs):
 
         if self.confustion_matrix is None:
-            self.confustion_matrix = multilabel_confusion_matrix(labels, outputs, labels=[0, 1])
+            self.confustion_matrix = multilabel_confusion_matrix(labels, outputs, labels=[0, 1]).astype(np.float32)
 
         else:
-            self.confustion_matrix += multilabel_confusion_matrix(labels, outputs, labels=[0, 1])
+            self.confustion_matrix += multilabel_confusion_matrix(labels, outputs, labels=[0, 1]).astype(np.float32)
 
     def compute(self):
         f1 = 0
